fix: try 5

+++ 0x1E-search_algorithms/100-jump.c
@@ -17,24 +17,24 @@
 
 int jump_search(int *array, size_t size, int value)
 {
-	int step = sqrt(size);
-	int start = 0, end = 0;
+	size_t step = sqrt(size);
+	size_t start = 0, end = 0;
 
 	if (!array)
 		return (-1);
 
-	while (value > array[end] && end < (int)size)
+	while (value > array[end] && end < size)
 	{
-		printf("Value checked array[%d] = [%d]\n", end, array[end]);
+		printf("Value checked array[%ld] = [%d]\n", end, array[end]);
 		start = end;
 		end += step;
 	}
 
-	printf("Value found between indexes [%d] and [%d]\n", start, end);
+	printf("Value found between indexes [%ld] and [%ld]\n", start, end);
 
-	while (start <= end && start < (int)size)
+	while (start <= end && start < size)
 	{
-		printf("Value checked array[%d] = [%d]\n", start, array[start]);
+		printf("Value checked array[%ld] = [%d]\n", start, array[start]);
 		if (array[start] == value)
 			return (start);
 		start++;
